Moved categorize_weights to font_properties module
